# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Gestionale Sagra API
  version: 1.0.0
  description: API REST per la gestione di categorie, cibi e ordini in una sagra
servers:
  - url: http://localhost:3030
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Autenticazione utente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "password"
      responses:
        200:
          description: Login effettuato con successo
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "token123"
        401:
          description: Credenziali non valide
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...